refactor: HoneyTipDefaultRepository -> HoneyTipRepository 로 이동 (#230)
