+++ src/main/java/com/api/ttoklip/domain/honeytip/post/repository/HoneyTipRepositoryImpl.java
@@ -15,7 +15,9 @@ import com.api.ttoklip.domain.honeytip.post.domain.HoneyTip;
 import com.api.ttoklip.global.exception.ApiException;
 import com.api.ttoklip.global.exception.ErrorType;
 import com.querydsl.core.types.dsl.BooleanExpression;
+import com.querydsl.core.types.dsl.NumberExpression;
 import com.querydsl.core.types.dsl.Wildcard;
+import com.querydsl.jpa.impl.JPAQuery;
 import com.querydsl.jpa.impl.JPAQueryFactory;
 import java.util.List;
 import java.util.Optional;
@@ -138,6 +140,98 @@ public class HoneyTipRepositoryImpl implements HoneyTipRepositoryCustom {
                 .orderBy(honeyTip.id.desc())
                 .limit(3)
                 .fetch();
+    }
+
+    @Override
+    public List<HoneyTip> getHouseWork() {
+        JPAQuery<HoneyTip> query = defaultQuery();
+
+        return query
+                .where(
+                        honeyTip.category.eq(Category.HOUSEWORK),
+                        getHoneyTipActivate()
+                )
+                .fetch();
+    }
+
+    private JPAQuery<HoneyTip> defaultQuery() {
+        return jpaQueryFactory
+                .selectFrom(honeyTip)
+                .distinct()
+                .leftJoin(honeyTip.honeyTipComments, honeyTipComment)
+                .leftJoin(honeyTip.honeyTipLikes, honeyTipLike)
+                .leftJoin(honeyTip.honeyTipScraps, honeyTipScrap)
+                .limit(10)
+                .orderBy(honeyTip.id.desc());
+    }
+
+    @Override
+    public List<HoneyTip> getRecipe() {
+        JPAQuery<HoneyTip> query = defaultQuery();
+
+        return query
+                .where(
+                        honeyTip.category.eq(Category.RECIPE),
+                        getHoneyTipActivate()
+                )
+                .fetch();
+    }
+
+    @Override
+    public List<HoneyTip> getSafeLiving() {
+        JPAQuery<HoneyTip> query = defaultQuery();
+
+        return query
+                .where(
+                        honeyTip.category.eq(Category.SAFE_LIVING),
+                        getHoneyTipActivate()
+                )
+                .fetch();
+    }
+
+    @Override
+    public List<HoneyTip> getWelfarePolicy() {
+        JPAQuery<HoneyTip> query = defaultQuery();
+        return query
+                .where(
+                        honeyTip.category.eq(Category.WELFARE_POLICY),
+                        getHoneyTipActivate()
+                )
+                .fetch();
+    }
+
+    @Override
+    public List<HoneyTip> getTop5() {
+        return jpaQueryFactory
+                .selectFrom(honeyTip)
+                .distinct()
+                .leftJoin(honeyTip.honeyTipComments, honeyTipComment)
+                .leftJoin(honeyTip.honeyTipLikes, honeyTipLike)
+                .leftJoin(honeyTip.honeyTipScraps, honeyTipScrap)
+                .where(
+                        getHoneyTipActivate()
+                )
+                .groupBy(honeyTip.id)
+                .orderBy(
+                        getLikeSize()
+                                .add(getCommentSize())
+                                .add(getScrapSize())
+                                .desc()
+                )
+                .limit(5)
+                .fetch();
+    }
+
+    private NumberExpression<Integer> getScrapSize() {
+        return honeyTip.honeyTipScraps.size();
+    }
+
+    private NumberExpression<Integer> getCommentSize() {
+        return honeyTip.honeyTipComments.size();
+    }
+
+    private NumberExpression<Integer> getLikeSize() {
+        return honeyTip.honeyTipLikes.size();
     }
 
 }
